java/src/test/java/com/studio3104/leetcode/explore/collection/medium/DP/JumpGame/SolutionTest.java
package com.studio3104.leetcode.explore.collection.medium.DP.JumpGame;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

class SolutionTest {
    Solution solution = new Solution();

    static Stream<Arguments> argumentsProvider() {
        return Stream.of(
                Arguments.of(new int[]{2, 3, 1, 1, 4}, true),
                Arguments.of(new int[]{3, 2, 1, 0, 4}, false)
        );
    }

    @ParameterizedTest
    @MethodSource("argumentsProvider")
    void canJump(int[] nums, boolean expected) {
        boolean actual = solution.canJump(nums);
        Assertions.assertEquals(expected, actual);
    }
}
